requirements.txt
numpy>=1.24
pandas>=2.0
requests>=2.31
tqdm>=4.66
beautifulsoup4>=4.12
spacy>=3.7
transformers>=4.37
sentence-transformers>=2.5
faiss-cpu>=1.7
langchain>=0.1
langchain-community>=0.0.32
langchain-huggingface>=0.0.8
langchain-groq>=0.0.1
bs4>=0.0.2
